Back Propagation Working

+++ NeuralNetwork/Trainning/GA.h
@@ -12,7 +12,7 @@ namespace GA
     void Train(
         std::vector<std::vector<double> >       train_data,
         uint32_t                                population,
-        uint32_t                                mutation_amount,
+        double                                  mutation_amount,
         uint32_t                                generations_amount,
         BaseNetwork*                            base_network
     );
